+++ src/Unpacker/ProductPriceUnpacker.php
@@ -6,8 +6,9 @@ namespace Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Unpacker;
 use Heptacom\HeptaConnect\Dataset\Ecommerce\Currency\Currency;
 use Heptacom\HeptaConnect\Dataset\Ecommerce\Price\Condition;
 use Heptacom\HeptaConnect\Dataset\Ecommerce\Price\Price;
-use Heptacom\HeptaConnect\Dataset\Ecommerce\Product\Product;
+use Heptacom\HeptaConnect\Dataset\Ecommerce\Price\PriceCollection;
 use Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Support\DalAccess;
+use Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Support\DalSyncer;
 use Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Support\ExistingIdentifierCache;
 use Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Support\PrimaryKeyGenerator;
 use Ramsey\Uuid\Uuid;
@@ -39,9 +40,32 @@ class ProductPriceUnpacker
         $this->priceConditionUnpacker = $priceConditionUnpacker;
     }
 
-    public function unpack(Price $price, Product $product): array
+    public function unpack(PriceCollection $sourceCollection, string $productNumber): iterable
     {
-        $ruleId = $this->preparePriceRuleId($price);
+        $syncer = $this->dalAccess->createSyncer();
+
+        $ruleIds = [];
+
+        /** @var Price $sourcePrice */
+        foreach ($sourceCollection as $key => $sourcePrice) {
+            $ruleIds[$key] = $this->preparePriceRuleId($sourcePrice, $syncer);
+        }
+
+        if ($syncer->getOperations() !== []) {
+            $syncer->flush();
+        }
+
+        /** @var Price $sourcePrice */
+        foreach ($sourceCollection as $key => $sourcePrice) {
+            /** @var string $ruleId */
+            $ruleId = $ruleIds[$key];
+
+            yield $this->unpackProductPrice($sourcePrice, $productNumber, $ruleId);
+        }
+    }
+
+    protected function unpackProductPrice(Price $price, string $productNumber, string $ruleId): array
+    {
         $priceId = PrimaryKeyGenerator::generatePrimaryKey(
                 $price,
                 'da210b7c-fd7c-4aa6-a0ee-846a508482db'
@@ -50,7 +74,7 @@ class ProductPriceUnpacker
         $price->setPrimaryKey($priceId);
         $price->setPrimaryKey($price->getPrimaryKey() ?? Uuid::uuid5(
                 $ruleId,
-                $product->getNumber().'__'.$price->getQuantityStart()
+                $productNumber.'__'.$price->getQuantityStart()
             )->getHex());
 
         if ($price->getCurrency() instanceof Currency) {
@@ -97,20 +121,20 @@ class ProductPriceUnpacker
         return $targetPrice;
     }
 
-    protected function preparePriceRuleId(Price $sourcePrice): string
+    protected function preparePriceRuleId(Price $sourcePrice, DalSyncer $syncer): string
     {
         $targetConditions = [];
         $nameParts = [];
 
         $targetConditions[] = $orMergeCondition = [
-            'id' => static fn(string $ruleId): string => Uuid::uuid5(self::NS_CONDITION_CONTAINER_OR, $ruleId)->getHex(),
+            'id' => static fn(string $ruleId): string => (string) Uuid::uuid5(self::NS_CONDITION_CONTAINER_OR, $ruleId)->getHex(),
             'type' => 'orContainer',
             'position' => 0,
             'value' => [],
         ];
 
         $targetConditions[] = $andMergeCondition = [
-            'id' => static fn(string $ruleId): string => Uuid::uuid5(self::NS_CONDITION_CONTAINER_AND, $ruleId)->getHex(),
+            'id' => static fn(string $ruleId): string => (string) Uuid::uuid5(self::NS_CONDITION_CONTAINER_AND, $ruleId)->getHex(),
             'parentId' => $orMergeCondition['id'],
             'type' => 'andContainer',
             'position' => 0,
@@ -151,7 +175,7 @@ class ProductPriceUnpacker
 
         \usort($conditionEssences, static fn (array $a, array $b): int => \json_encode($a) <=> \json_encode($b));
 
-        $ruleId = Uuid::uuid5('a7a0d619-3fc3-40f6-b57c-31f895ae652a', \json_encode($conditionEssences))->getHex();
+        $ruleId = (string) Uuid::uuid5('a7a0d619-3fc3-40f6-b57c-31f895ae652a', \json_encode($conditionEssences))->getHex();
 
         foreach ($targetConditions as &$targetCondition) {
             $targetCondition['id'] = \is_callable($targetCondition['id']) ? $targetCondition['id']($ruleId) : $targetCondition['id'];
@@ -172,7 +196,9 @@ class ProductPriceUnpacker
             'conditions' => $targetConditions,
         ];
 
-        $this->dalAccess->repository('rule')->upsert([$rule], $this->dalAccess->getContext());
+        if (!$this->dalAccess->idExists('rule', $ruleId)) {
+            $syncer->upsert('rule', [$rule]);
+        }
 
         return $ruleId;
     }